Revert "Add more SdkResolverService events and allow SdkResolvers to log events (#7139)" (#7277)
This reverts commit 73ee6c0b41ebfa80f5dd54560a90566a8efb20cb.

### Description

As of #7139 Microsoft-Build events can no longer be reported (i.e. they don't show in PerfView) because the ETW manifest builder is failing due to Object[] being an unsupported type. This change is a straight revert of the PR that introduced the regression.

### Customer Impact

The bug makes it harder to diagnose MSBuild issues internally and externally. Note that Microsoft-Build ETW events are recorded by the VS Feedback tool, for example.

### Regression?

Yes, introduced in #7139 on Dec 22nd 2021.

### Risk

Very low.

### Is there a packaging impact?

No.

### Does the change affect files included in any ref pack (Microsoft.NETCore.App.Ref, Microsoft.AspNetCore.App.Ref, Microsoft.WindowsDesktop.App.Ref)?

No.