Default loggers ProjectConfigurationDescription
dotnet/sdk#12030 discovered that the .NET SDK's default build process
didn't log TargetFramework details for normal projects.

This turned out to be because the .NET SDK attaches a custom logger,
which caused MSBuild to set up a ConfigurableForwardingLogger in front
of the ConsoleLogger that forwarded based on verbosity which wasn't high
enough to include the ProjectEvaluationFinished event.

Pass a new ConfigurableForwardingLogger parameter from the command line
handler to instruct it to pass those events even at lower verbosities.