Remove STANDALONEBUILD
This was used during the initial open-sourcing of MSBuild, when it was mirrored
between GitHub and the internal VS repo. STANDALONEBUILD was defined only for the
GitHub copy.

The internal VS mirror of MSBuild has been deleted since 15.0, and as far as I
know no one has complained about any of these things being missing, so remove the
ifdefs.