Check wow6432 registry node for clsid registration in comimporter clickonce sources (#7328)
ClickOnce's msbuild task to generate the application manifest has code that imports COM Component by reading their registration data from the registry. In this scenario, the COM reference is an x86 binary that is registered under HKCR\WOW6432Node\Clsid node. In dev17 with msbuild being an x64  process, the code is not reading from the WOW6432Node.

The fix is to update the code that's reads COM registration data to look under the WOW6432 node of HKCR as well.

Fixes https://dev.azure.com/devdiv/DevDiv/_workitems/edit/1463446.

Co-authored-by: Sujit Nayak <sujitn@exchange.microsoft.com>