Improved error message when created node crashes (#7306)
Occasionally, someone will have a misconfiguration that causes MSBuild
worker nodes to crash immediately or shortly after launch (before
establishing communication with the node that created them).

This can be really hard to diagnose, because it's not obvious that it's
happening--failure to connect to a process is nonfatal and can happen
if a user is building with two disconnected `-m` invocations. But
repeated failure is bad and if the node _never_ connects, that's a
problem.

Add comm-trace information about process lifetime. I wanted to include
the exit code too but couldn't because on Windows we call CreateProcess
instead of Process.Start.