Implement warnnotaserror Fixes #3062 (#7309)
Fixes #3062

Context
We previously had warnaserror (and a property for it) that, when you specified error codes, upgraded those error codes from warnings to errors. If you just left it empty, it upgraded all warnings to errors. (Null meant don't upgrade.)

This adds that you can ask for all error codes to be upgraded, then downgrade just a few of them (via codes) back to warnings.

Changes Made
Implement WarnNotAsError both as a command line switch and as a property.

Testing
Added a unit test. Tried it out from the command line.