Provide nice error when an environment variable is not expanded on the command line, including the full command line for all switch errors Fixes #7210 (#7213)
In addition to fixing the error below, this captures information about the command line (including parts that came from .rsp files) and logs those along with any command line switch error. For switches from response files, it lists where they came from.

Fixes #7210

Context
When you specify an environment variable on the command line or via Directory.Build.rsp, and that environment variable is not defined, it puts the literal in instead. MSBuild misinterprets that as a duplicate project file and fails. This works around that.

Changes Made
Ignore environment variable "projects" if another project is found. (Otherwise doesn't check.)

Testing
Tried with a simplified repro. Fixed the problem.