Expand architecture customization for usingtasks
Use RuntimeInformation.ProcessArchitecture in GetCurrentMSBuildArchitecture.

Refactor ToolsDirectory computation

BuildEnvironmentHelper goes through a complicated process to compute the
paths to the 32-bit and 64-bit tools folders.

Refactored the logic to expose a new ToolsDirectoryRoot that is "the
base folder" and simplify construction of the more-specific ones to
append to the root.

Don't distinguish tools directory root in handshake

Node communication improvements