Simplify MainNodeSdkResolverService (#7244)
Fixes #7137

Context
The MainNodeSdkResolverService process incoming requests from out-of-proc nodes to handle SDK resolution. The main node has a queue of all incoming packets that are processed by appropriate handlers. The MainNodeSdkResolverService currently takes SdkResolverRequests and places them in a queue for processing. This seems unnecessary to have two queues and two threads draining the queues.

Changes Made
Remove queue and thread that process it in favor of launching a fire-and-forget task that resolves the SDK and sends the response.

Testing
Existing tests cover this area