Fix missing transitive dependencies (#7476)
In #7053 there's a subtle bug which is highly dependent on node ordering (which makes it even repro differently or not at all on different machines...).

Specifically, `traversedReferences` is "reset" every time `GetTransitiveProjectReferencesExcludingSelf` is called. So we can end up in a state where a node is in the `transitiveReferenceCache` but not in `traversedReferences`. In that case, `GetTransitiveProjectReferencesExcludingSelfHelper` will add the cached transitive references to the `traversedReferences` set, but *not* `incompleteReferencesOfDirectlyReferencingNode`, so the referencing node loses these transitive references.

I reworked the `GetTransitiveProjectReferencesExcludingSelf` method to a more straightforward approach. First it checks the cache. Then it creates a new set for the transitive references for this node and immediately adds it to the cache (note the set will be mutated later). This helps cycle handling (this avoids the StackOverflow from #6925, but does make the transitive references incorrect in the cycle case, but that's fine since we'll error later anyway when doing proper cycle detection). Then just iterate the references, add them and add the recursion into their transitive references.

Fixes #7475