Allow overriding MSBuildRuntimeType to "Full" on RUNTIME_TYPE_NETCORE (#7481)
Cause MSBuild to set `MSBuildRuntimeType=\"Full\"` when `MsBuildForceEvaluateAsFullFramework` is set.

We have scenarios in internal build tools where MSBuild evaluation is executed in the parse phase under a .NET 6.0 process, prior the build phase which builds projects using full framework `msbuild.exe`. As evaluation results in the parse phase must match the build phase for prediction to be accurate, `MSBuildRuntimeType` must be set to `\"Full\"` in both cases - so we currently solve this by launching a separate process under full framework which runs evaluation / parsing and serializes results back to the main process.

However, this serialization incurs a significant amount of overhead on large repos. Instead, by having the option of forcing `MSBuildRuntimeType=Full` when on `RUNTIME_TYPE_NETCORE`, we would be able to evaluate these projects in-process for the majority of repos.