Ensure path writable for MSBuildDebugEngine (#7366)
If `MSBUILDDEBUGPATH` is invalid, it will now attempt to redirect to the current working directory. If writing to the current working directory fails, perhaps because it's in `C:\Program Files` or another protected directory, it redirects to `TEMP`. If the path is too long, it now fails with a nice error.

Fixes #7241