Remove opt in for new schema for CombineTargetFrameworkInfoProperties (#6928)
* Remove opt in for new schema

For CombineTargetFrameworkInfoProperties, we had hit an issue in which a TF could be invalid as an XML root. We changed the schema to let us escape it, then opted into the new schema in the SDK. This removes the option to not opt into the new schema. It's hopefully still ok to do this; though it would have been better to get it into 17.0, I don't think anyone other than the SDK is using this, so I don't think it should matter.

* Update src/Tasks/Microsoft.Common.CurrentVersion.targets

* Put back attribute

* Put in change wave

* Update src/Tasks/CombineTargetFrameworkInfoProperties.cs