Parallelize Node Spawning (#7513)
Fixes #7379

Context
MSBuild builds with Out Of Proc nodes which requires spawning processes. Each node takes time to spawn and is serial. This can lead to a slow start on machine with many cores.

Changes Made
Parallelize Node spawn.
Changed interface in INodeProvider to specify number of nodes to be created and let particular implementation of INodeProvider decide if it does it concurrently. This way the scope of thread safety is limited and manageable.

I tried to keep current logic as much in tact as possible. For example old logic allowed to call INodeManager to spawn multiple in-proc nodes but then it fails anyway. That does not make much sense to me, and if we want to address it please let me know.

Testing
Various manual testing including scenarios with TaskHost.

Notes
When nodereuse and nodes are still running the difference is not significant about 20ms per node as connecting to existing processes is quite fast.
In cold scenarios, either /nr=false or nodes have had terminated after 15 minutes of idling, the difference is 150ms per node.
So on my 24 cores machine, incremental building of orchard code in cold scenario were about 20% faster, in hot scenarios (reusing still running nodes) about 4% faster.