Compile AppDomainIsolatedTask for all targets (#7589)
Today, if you want to run an `AppDomainIsolatedTask` in `dotnet build`,
it fails because `AppDomainIsolatedTask` can't be identified. With this
change, a user can make the attempt to run the task, for instance with
TaskHost isolation instead.

This was extra silly, because the main thing `AppDomainIsolatedTask`
does is apply the `[LoadInSeparateAppDomain]` attribute, which _was_
already working in .NET Core MSBuild.