Platform Negotiation: Platform is considered a referenced project's DEFAULT platform (#7511)
Reduces the number of evaluations in platform-negotiated builds.

Context
A (Platform=x86) -> B (Platforms=x64;x86, Platform=x86) (A, a project built as x86 references B, a project that can build as x86 or x64, but will default to x86 if built in isolation)

Prior to this change, platform negotiation would build B with SetPlatform=Platform=x86, causing a new evaluation. This can be avoided because it would have built as x86 with or without that global property being passed.

Changes Made
GetTargetFrameworks target now also returns the Platform property from the target. During platform negotiation, if the referenced project's platform matches the current project's platform, we now decide to build that project without passing a global Platform property. This prevents a new evaluation and will maximize the number of re-used evaluations during platform negotiation.

Testing
Created new UT for this.