Add location string to drive enumerating wildcard warning (#7553)
Fixes #7029 by associating a file location string with the WildcardResultsInDriveEnumeration resource name.

Context
For https://github.com/dotnet/project-system/blob/main/build/import/Workarounds.targets#L8, it was difficult to detect the location at which the drive enumerating wildcard pattern was occurring even though a warning was logged, since 1) there was no location string added to the warning, and 2) the Exclude rather than the Remove element was shown in the warning. To resolve 1), it would be better to associate the location string with the WildcardResultsInDriveEnumeration resource name.

Changes Made
Added location string to the WildcardResultsInDriveEnumeration resource name in order to view the location for drive enumerating wildcard warnings and errors.

Testing
Modified ProjectItem, ObjectModelHelper (for ProjectItemInstance), and CreateItem unit tests to ensure that the project file location was placed in the warning or error message.

Notes
Any <i Remove=<drive enumerating wildcard> /> will still be considered as an Exclude attribute in the warning or error message, since is treated as equivalent to <Foo Include=... Exclude=... />, which still makes the warning or error message slightly unclear for drive enumerating wildcard used in Remove cases. However, the wildcarded value will still be caught and logged/thrown based on whether the environment variable is set.