Version Microsoft.DotNet.Compatibility same as SDK (#25590)
We discussed offline that it makes sense to version this component alongside the SDK. We just recently noticed that during servicing, this package causes conflicts as the same version already exists on NuGet. Versioning it alongside the SDK will fix this during servicing and makes the component more consistent with the rest of the shipping assets.