Compatibility tooling refactoring (#25765)
* Update solution files and add Console proj

* Move logger into ApiCompatibility and extend them

Move the loggers into ApiCompatibility as logging isn't
PackageValidation specific. This allows to expose the ApiCompatibility
as an task be able to use logging and generate the suppression
file.

Update the resource files becaues of the move.

Change the ICompatibilityLogger type from an interface to an abstract
class so that it can serve a default implementation for generating the
suppression file. Also change the return type of the
LogError/LogWarning methods to boolean so that callees know if the
message was actually logged.

* MetadataInformation needs to override both Equals overloads

* Move SuppressionEngine into Logging namespace

Move the SupressionEngine and Suppression types into the Logging
namespace.

Make Suppression.DiagnosticId a required field (via nullability compiler
checks) and simplify the construction of the SuppressionEngine.

* Nit, comment fix

* Simplify Package type creation and code clean-up

Merge NupkgParser and Package types and make the instantiation of the
type more performant by avoiding Linq re-enumerations.

Also enable nullability for the type.

* Nit msbuild code clean-up

* Update files for nullability and optimize code-paths

* Update validators for nullability and other refactoring

* nit, test namespace fixes and using updates

* Remaining test updates

* Update MetadataInformation.cs

* Update Microsoft.DotNet.Compatibility.csproj

* Use System.CommandLine for console project

* Simplify Equals check syntax

* Clean-up logger parameters

* Move common frontend code into Shared, Validators

Move common frontend code (initialization and execution of the
validators) into a shared project. Update solution files.

Add some properties to make the console project packable and consumeable
as a global tool.

Update the design of the validators and define an interface so that they
can work with multiple packages passed in via multiple Validate method
invocations.

* Update tests

* Fix frameworkreferences null handling in ApiCompatRunner

* Add StrongNameKeyId Open for correctness