Fallback to 32/64 bitness logic for unknown architectures (#7731)
Previously, MSBuild checked the architecture by looking at the size of an IntPtr. With the changes needed for arm64, we threw an exception for any not-well-known architectures. This breaks scenarios that previously worked (arch: s390x).

Stop throwing an exception for unknown architectures. Instead, fall back to the logic that previously worked as a "best effort"

Fixes #7729