add rules to catch addition or removal of 'virtual' keyword (#25734)
This change emits a diagnostic when the virtual keyword is removed.
In strict mode, it emits a diagnostic when the virtual keyword is added.

This is done to reduce the number of false positives for what has been
treated as a backwards-compatible change in the past. We intentionally
skip interface members, since it is unclear how default implementation of
interface methods and sealed interface methods play into this. See #26169.

Fixes #18682
Updates #18683