Avoid locking TaskHostFactory-using assemblies  (#7387)
Getting type information for a task that was supposed to be loaded in a TaskHost involves loading that assembly today. The point of TaskHost nodes is that they don't lock the assembly, and they disappear quickly, releasing the lock from when we actually need to load the assembly to execute it. This will mean that we don't lock the assembly to read its type information, actually complying with the user's likely intent.

If the user asks for a TaskHost, use MetadataLoadContext to get its information rather than loading the assembly fully.

Fixes #6461