Make folder if needed for MSBUILDDEBUGPATH (#7791)
When using MSBuildDebugEngine, we recommend specifying MSBUILDDEBUGPATH. If you specify somewhere we can't write to, we have fallbacks. However, if you specify somewhere we can write to but doesn't yet exist, we can't write to it yet, so that part returns false, and we go on to the fallbacks. This makes the folder explicitly if we can, making it return true if we can write to a directory, but it doesn't yet exist.

Fixes #7752