Check derivable from ITaskItem directly (#7810)
Previously checked if any base type was named ITaskItem

Context
A type derived from a MetadataLoadContext does not inherit from any type not derived from the same MetadataLoadContext. To work around that, the previous code checked whether each input passed to a task type can be cast to an ITaskItem by comparing it and its base types to ITaskItem by full name. There may be multiple plausible base types, however. This corrects that by checking if ITaskItem IsAssignableFrom the type, ensuring that a separate "ITaskItem" is passed for MLC and non-MLC contexts.

Customer Impact
Customers with custom tasks that take in or return types derived from ITaskItem (or ITaskItem[]) rather than ITaskItem (or ITaskItem[]) itself cannot build.

Regression?
Yes, in 17.3p3.

Risk
I believe it is reasonably low, but I would like @rainersigwald to comment.

Changes Made
Use IsAssignableFrom instead of checking by name.

Testing
Tried building a simple test project, and it worked as expected. Tried building a more complicated test project that used TaskFactory="TaskHostFactory" in a UsingTask with a custom Task assembly, and it worked. Tried building the sln that had failed leading to this issue, and it built successfully.