Simplify exception handling for graph builds (#7831)
Just some cleanup and simplification in the exception handling for graph builds.

Functionally here are the differences:

ParallelWorkSet now throws an aggregate exception containing all exceptions instead of just the first exception that gets thrown.
ParallelWorkSet tasks no longer go away when they throw. They're caught and added to a list used in the point above.
BuildManager now properly logs invalid project errors for each invalid project (due to the help of the points above). Previously it was attempting to catch the agg exception which was never thrown and also the way it was in the catch effectively made it unreachable in the first place.
GraphBuildResult now exposes the CircularDependencyException via the Exception field instead of Exception being null while CircularDependency was true.