SetPlatform Negotiation: No global properties during GetTargetFrameworks (#7803)
Fixes #7760

Context
#7511 figured out a better way to prevent over-evaluations when A(Platform=x86) -> B(Platform=x86, Platforms=x64,x86). It allowed setplatform negotiation to check "would project B have built as what A is currently building as if we didn't tell it anything?

Unfortunately, there's a bug when passing a global property because GetTargetFrameworks passes Platform and Configuration, despite not needing to. This PR is an attempt at resolving this by no longer passing those properties, as well as undefining them.

Changes Made
Remove additional properties during the MSBuild call to GetTargetFrameworks.

Testing
Notes
It's possible we can't fix it this way, and instead we'll need to create two msbuild calls. One that's the standard but only gets called when EnableDynamicPlatformResolution is false, and the other that only gets called when EnableDynamicPlatformResolution is true.