Add baseline package validation strict mode (#26926)
Fixes https://github.com/dotnet/sdk/issues/25302

Adds a strict mode option for baseline validation to the apicompat
frontends. This mode is intended to be used during servicing, when in a
patch version update, no new public api additions are allowed.

Adds two tests that verify that the switch works when enabled and
disabled.