Support left and right names for assembly set (#27024)
* Support left and right names for assembly set

The existing logic allows to specify a left's name and multiple rights'
names. When comparing an assembly set against one or multiple other
assembly sets, the left and right names are different (for each
assembly).

To support multi assembly mode (assembly set comparison), remove the
hardcoded left and right names values in the runner and instead infer
the name from the mapped (containing) assembly. To faciltate that,
create a hierarchy in the element mapper types.