Support custom rule settings, improve testability (#27138)
Declaring a IRuleFactory interfaces that is responsible for providing
rules to run. Provide the RuleFactory implementation by the frontend so
that they can pass custom settings to the individual rules
(i.e. for exclude attributes files).

Improve testability of types like ApiComparer and RuleRunner by
depending on interfaces only instead of concrete types. That allows to
fully unit test these types.

Updating the rules to depend on an interface instead of the concrete
type for the rule context.

Updating the tests as passing settings to ApiComparer changed as the
type itself is now immutable.