Support loading assembly references when multiple assemblies are passed to the AssemblySymbolLoader (#27237)
* Support multiple assemblies in the SymbolLoader

Today, the assembly symbol loader does allow to load multiple
assemblies, but when one of them references another one of the passed in
assemblies, the reference fails to resolve when that assembly isn't also
passed in as an external reference (search path directory / search path
file).

Instead we want the symbol loader to load all passed in assemblies and
only resolve the references after just the assemblies got resolved.

In addition, by adding the functionality to directly load assemblies
from an archive (nuget package), the archive stream needs to only be
opened once instead of n-times, per "LoadAssembly" call.
Deleting the MetadataStreamProvider as it isn't necessary anymore with
the "LoadAssembliesFromArchive" change.