Use Logger and apply fix in AssemblyIdentity rule (#27233)
* Use Logger and apply fix in AssemblyIdentity rule

The AssemblyIdentityRule did incorrectly emit a difference when a right
assembly is passed in without a matching left in non strict mode. Non
strict mode should allow passing in more on the right hand side than on
the left side.

In the case of when a single assembly is compared, the difference should
be emitted as a missing left indicates a wrong user input.

Also passing in the logger to the rule in order to still log a
diagnostic message in non strict mode indicating that the assembly couldn't be found.