Fix roslyn assemblies path not being respected (#27256)
* Fix roslyn assemblies path not being respected

In the tools package, the roslyn assemblies path wasn't respected
because it was always resolved from the app local path. By moving the
codeanalysis files out into a subdirectory, a different version can be
loaded instead. This is using the same approach as the msbuild task
which also needs to manually resolve the codeanalysis version.

Also fixing the default verbosity level of the global tool.

* Setting IsShippingPackage true