Update binding redirect for System.Reflection.Metadata to match referenced assembly version
Update the binding redirect for System.Reflection.Metadata to match the referenced version of the assembly from the current package reference.

When this is out of date, it causes a lot of pain in consuming repos (in particular in dotnet/sdk and dotnet/runtime). It might be worth some tooling to ensure that this stays up to date.