Add new output DebugSymbolsFiles to ResolvePackageAssets
DebugSymbolsFiles outputs the list of absolute path to symbols files
for package references.
ReferenceDocumentationFiles outputs the list of absolute path to Xml
documentation files.

The change:
ResolvePackagesAssets was modified to read/write DebugSymbolsFiles
and ReferenceDocumentationFiles from/to the cache file. They were inserted
after ContentFilesToProcess to keep the alphabetical order.

The target was modified to copy the DebugSymbolsFiles to output
directory only when CopyDebugSymbolFilesFromPackages is true, and
ReferenceDocumentationFiles to the output directory when
EnableCopySymbolsAndXmlDocsToOutputDir is true.

There is some minor code cleanup not related to this bug.