[Blazor] Eliminate duplicate inputs to ComputeBlazorBuildAssets (#27561)
There should be no duplicates in ReferenceCopyLocalPaths, but when a bug
causes it to be, it creates issues for Blazor. This change makes our
pipeline more resilient against these issues by removing the duplicates
and giving the build a chance to succeed in spite of bugs that introduce
duplicate items.