Merged PR 423224: Update 16.7 build definition
This PR had two main parts: modernizing the .vsts-dotnet.yml file in our repo and upgrading arcade.

For the first, we had 4 main parts. This updated the pool we used for our build to be something that exists, switched to using a personal access token because our prior authentication method had been removed, stopped using a couple no-longer-functional azure devops tasks (opting for modern versions), and set our vm image to be one we have access to—i.e., a modern one.

For the second, we waffled slightly on the arcade version to update to but ultimately switched to version 1.0.0-beta.21065.4 of arcade (6b9ad1b86345fb7fb81899b230fed387c611d55a). On top of that, we needed a slight tweak to the .net cli version, as we'd gotten off there.