Don't throw if a temp file already exists
When using TempFileUtilities.CreateTempFile, it throws if the file already exists. For 16.7, we expanded its usage to include creating the "file" to be used by a unix socket. That file should not be created manually because it needs to be created a socket. With the assert, we noticed the socket-file and threw erroneously.