Reduce unnecessary allocations when accessing information about global properties (#7887)
Fixes #7743

CPS iterates over collections of properties to access information about them, specifically whether they are global properties or derived from the environment. The current implementation of IsGlobalProperty and IsEnvironmentProperty start by accessing the global dictionary. That call makes a copy of the dictionary before just accessing information, which impacts performance.

This gets more complicated because the global dictionary should not be modified from the outside. Based on our publicly exposed API, it should be an IDictionary, so we cannot return an ImmutableDictionary. Returning a readonly Dictionary still permits modifications, and a ReadOnlyDictionary doesn't implement IDictionary.

This attempts to circumvent the problem of returning a readonly dictionary that implements IDictionary by instead exposing a more efficient form of the few methods needed by CPS, specifically Contains, Count, and Enumerate. As CPS uses a custom ProjectLink, these new methods are exposed as overridable methods, and they're virtual to avoid making a breaking change. This permits performance wins without breaking current customers.